% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateRegressionModel.R
\name{evaluateRegressionModel}
\alias{evaluateRegressionModel}
\title{Evaluate regression (continuous output prediction) models.}
\usage{
evaluateRegressionModel(time, actual, predicted, title = NULL,
  lim_val = NULL, calcMAPE = TRUE, scatterplot = TRUE,
  trendplot = TRUE, residualplot = TRUE, scale_attributes = NULL)
}
\arguments{
\item{time}{Vector of DateTime (POSIXct) to be used as index for plotting.}

\item{actual}{Actual values of the parameter being predicted (y).}

\item{predicted}{Predicted values of the parameter being predicted (y-hat).}

\item{title}{Title to be added to the plots. Defaults to \emph{NULL} with no title included.}

\item{lim_val}{Single numeric value as limit value for plotting (note for ggplot2 objects lims() is used).
This does not affect the evaluation metric calculations.}

\item{calcMAPE}{\emph{TRUE/FALSE} to indicate whether MAPE should be calculated or not. MAPE is not calculatable if there are 0's in the
actual values (results in \emph{Inf} due to division by 0). Defaults to \emph{TRUE}.}

\item{scatterplot}{\emph{TRUE/FALSE}, whether Actual vs Predicted scatterplot should be shown.}

\item{trendplot}{\emph{TRUE/FALSE}, whether Trend Plot of Actual & Predicted should be shown.}

\item{residualplot}{\emph{TRUE/FALSE}, whether histogram of residuals (Predicted - Actual) should be shown.}

\item{scale_attributes}{Dataframe object that details the Mean and Std_Dev of variables obtained from \code{scale_data()}.}
}
\value{
None
}
\description{
Evaluate regression (continuous output prediction) model outputs based on predictions and actual values.
This version only allows evaluation on temporal data (timestamp-based data) with a DateTime index.
For evaluation purposes, the MAPE, MAE, RMSE, and R2 values are returned.
}
\examples{
evaluateRegressionModel(data_df$DateTime, data_df$Actual, data_df$Predicted, title="Evaluation of Regression Model")

}
\seealso{
\code{\link{scale_data}}
}
