% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivePlot.R
\name{interactivePlot}
\alias{interactivePlot}
\title{Interactive Plotting}
\usage{
interactivePlot(data_df, regex_string = NULL,
  datetime_colname = "DateTime")
}
\arguments{
\item{data_df}{Dataframe for plotting. Must include a datetime column (POSIXct).}

\item{regex_string}{Regex-based searching for variables in \code{data_df}. Defaults to \emph{NULL}.}

\item{datetime_colname}{Column name for the datetime column in \code{data_df}. Defaults to "DateTime".}
}
\value{
None
}
\description{
Interactive trend plotting using \code{xts} and \code{dygraphs} packages in R for quick data exploration.
}
\details{
A GUI is shown for the user to select the variables to be plotted against a timeseries.
}
\examples{
interactivePlot(data_df)

}
