% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nalocf_partial.R
\name{nalocf_partial}
\alias{nalocf_partial}
\title{Partial Forward-Filling}
\usage{
nalocf_partial(x)
}
\arguments{
\item{x}{Vector object to be forward-filled until last observation.}
}
\value{
Vector of forward-filled data (until last observation).
}
\description{
Wrapper function around the \code{zoo::na.locf} function which only forward fills until last observation.
This is useful to prevent tail NAs from being forward filled.
}
\examples{
nalocf_partial(x)
data_df <- sapply(data_df, function(x) nalocf_partial(x))

}
