% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Improved Correlation Plotting}
\usage{
plot_correlation(correlation_matrix, order_type = "hclust",
  text_angle = 45, sig_level = 0.05, number_font_size = 1)
}
\arguments{
\item{correlation_matrix}{Correlation matrix obtained from \code{stats::cor()}.}

\item{order_type}{Character, the ordering method of the correlation matrix.
\itemize{
 \item "original" for original order.
 \item "AOE" for angular order of the eigenvectors.
 \item "FPC" for first principal component order.
 \item "hclust" for hierarchical clustering order (default).
 \item "alphabet" for alphabetical order.
}}

\item{text_angle}{Numeric, angle (degrees) for the text to be adjusted for better visualisation. Defaults to 45.}

\item{sig_level}{Numeric, significant level for p-value to be considered significant, whereby it would be coloured. Insignifcant
p-values would not be coloured. Defaults to 0.05.}

\item{number_font_size}{Numeric, font size. Defaults to 1.}
}
\value{
None
}
\description{
Improvised correlation plotting which includes ordering of variables, text angle adjustment, significnat level t-test, and
font size adjustment.
}
\details{
Runs using the \code{corrplot::corrplot()} function.
}
\examples{
cor_mat <- cor(data)
plot_correlation(cor_mat)

}
