% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_TrainTestSplit.R
\name{prepare_TrainTestSplit}
\alias{prepare_TrainTestSplit}
\title{Prepares Train-Test Data With Scaling}
\usage{
prepare_TrainTestSplit(data_df, type = NULL, complete.cases = TRUE,
  train_test_ratio = 0.7, train_validation_ratio = 0.8)
}
\arguments{
\item{data_df}{Complete dataset including both predictor/input variables (x) and prediction/output variable (y)}

\item{type}{String,
\itemize{
 \item "rescale" to rescale a new dataframe object based on the given \code{scale_attributes}.
 \item "unscale" to unscale a dataframe back to its original values.
 \item \emph{NULL} no scaling is done (Default).
}}

\item{complete.cases}{Logical, whether incomplete rows should be removed. Default is \emph{TRUE}.}

\item{train_test_ratio}{Ratio of (Train + Validation) Data to Test Data.}

\item{train_validation_ratio}{Ratio of Train Data to Validation Data.}
}
\value{
List of three objects:
\item{train_df}{Training dataframe}
\item{validation_df}{Validation dataframe}
\item{test_df}{Test dataframe}
}
\description{
Prepares train, validation, and test dataset with option to scale (Z-transform) parameters.
}
\details{
only numeric columns will be scaled, non-numeric columns are maintained as-is.
}
\examples{
traintestsplit_list <- prepare_TrainTestSplit(data_df, type="temporal")
train_df <- traintestsplit_list[[1]]
validation_df <- traintestsplit_list[[2]]
test_df <- traintestsplit_list[[3]]

}
\seealso{
\code{\link{rescale_data}} \code{\link{train_test_split}} \code{\link{scale_data}}
}
