% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_data.R
\name{rescale_data}
\alias{rescale_data}
\title{Rescaling/Unscaling Data}
\usage{
rescale_data(data_df, scale_attributes, scale_method)
}
\arguments{
\item{data_df}{Dataframe including both predictor/input variables (x) and prediction/output variable (y)}

\item{scale_attributes}{DataFrame of scaling attributes (Mean, Std Dev) which are obtained from \code{scale_data()} function}

\item{scale_method}{String,
\itemize{
 \item "rescale" to rescale a new dataframe object based on the given \code{scale_attributes}.
 \item "unscale" to unscale a dataframe back to its original values.
}}
}
\value{
Rescaled/unscaled dataframe.
}
\description{
Rescaling or unscaling of data.
}
\details{
Rescaling is generally used for rescaling of test data based on training data \code{scale_attributes}.

Unscaling is generally used for unscaling of predicted data for evaluation.
}
\examples{
data_df <- rescale_data(data_df, scale_attributes, "rescale")

}
\seealso{
\code{\link{scale_data}}
}
