% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Scale data (Z-Transformation)}
\usage{
scale_data(data_df)
}
\arguments{
\item{data_df}{Training dataset including both predictor/input variables (x) and prediction/output variable (y).}
}
\value{
List of two objects:
\item{scaled_data_df}{Scaled dataframe. Only detected numeric columns are scaled.}
\item{scale_attributes}{Dataframe of scale attributes (Mean, Std_Dev).}
}
\description{
Scales a dataframe based on Z-transformation (based on \code{base::scale()} function in R). Also returns a supplementary dataframe
which contains information on the Mean and Standard Deviation of the variables being scaled.
}
\details{
only numeric columns will be scaled, non-numeric columns are maintained as-is.
}
\examples{
scaled_data_list <- scale_data(data_df)
scaled_data_df <- scaled_data_list[[1]]
scale_attributes <- scaled_data_list[[2]]

}
\seealso{
\code{\link{rescale_data()}} \code{\link{train_test_split()}} \code{\link{prepare_TrainTestSplit()}}
}
