% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Train-Test Split}
\usage{
train_test_split(input_df, train_ratio, type)
}
\arguments{
\item{input_df}{Dataframe that is to be split.}

\item{train_ratio}{Ratio of training data to test data.}

\item{type}{String,
\itemize{
 \item "temporal" for time-dependant splitting of data.
 \item "random" for time-independent splitting of data (sampling at random).
}}
}
\value{
List consisting of:
\item{train_df}[Training dataframe]
\item{test_df}{Test dataframe}
}
\description{
Temporal-based or random-based train-test split.
}
\details{
This function can be used in sequence after another for Train-Validation-Test splitting.
}
\examples{
train_test_list <- train_test_split(data_df, 0.7, type="temporal")

}
\seealso{
\code{\link{prepare_TrainTestSplit}}
}
